Reverting example and renaming to `*.tsx`
